requirements.txt
requests
python-dotenv
praw
kafka-python
quixstreams
transformers
torch
duckdb
streamlit
plotly